assigning movement to eventdf
